refactor: rename file because it no longer export loader
