+++ lib/node/configuration.mjs
