+++ lib/node/mocha-loader.mjs
@@ -1,4 +1,4 @@
-import { configuration } from "./__common__.mjs";
+import { configuration } from "./configuration.mjs";
 import ValidateMocha from "../../dist/node/validate-mocha.mjs";
 import Mocha from "mocha";
 const { log } = configuration;
