+++ lib/node/recorder-mocha.mjs
@@ -1,4 +1,4 @@
-import { configuration } from "./__common__.mjs";
+import { configuration } from "./configuration.mjs";
 import RecorderMocha from "../../dist/node/recorder-mocha.mjs";
 
 const {
