+++ lib/node/recorder-process.mjs
@@ -1,4 +1,4 @@
-import { configuration } from "./__common__.mjs";
+import { configuration } from "./configuration.mjs";
 import RecorderProcess from "../../dist/node/recorder-process.mjs";
 
 const {
