+++ lib/node/recorder-remote.mjs
@@ -1,4 +1,4 @@
-import { configuration } from "./__common__.mjs";
+import { configuration } from "./configuration.mjs";
 import RecorderRemote from "../../dist/node/recorder-remote.mjs";
 
 const {
